Fix PFU references and update page title
Corrected template and URL name for "Power From Underground" to ensure consistency across the project. Updated the title in the associated HTML file for better clarity and context.

Started adding flash cycle.
Made the form to accept user input
Almost finished with the function view
Moving on to the html template now